% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Events}
\alias{Events}
\title{Events}
\format{
A data frame with 165 rows and 8 variables:
\describe{
  \item{ID}{Character string which refers to the identifier of the gauging station
      (in Austria: HZBCODE)}
  \item{EVENT_TYPE}{Event types are defined as follows:
    \itemize{
     \item 0: Constant event after NA event or constant event as first event
         in time series
     \item 1: Constant event after DC
     \item 2: Increasing event (IC)
     \item 3: Constant event after IC
     \item 4: Decreasing event (DC)
     \item 5: NA event}}
  \item{Time}{Date-time of event starting point}
  \item{AMP}{Amplitude (\code{\link[=amp]{amp()}})}
  \item{MAFR}{Maximum flow fluctuation rate (\code{\link[=mafr]{mafr()}})}
  \item{MEFR}{Mean flow fluctuation rate (\code{\link[=mefr]{mefr()}})}
  \item{DUR}{Duration (\code{\link[=dur]{dur()}})}
  \item{RATIO}{Flow ratio (\code{\link[=ratio]{ratio()}})}
}
}
\usage{
Events
}
\description{
A complementary dataset to the flow fluctuation dataset \link{Q}. It contains
the events and metrics such as computed by \code{\link[=get_events]{get_events()}}.
}
\keyword{datasets}
