% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Q}
\alias{Q}
\title{Flow Fluctuations Q}
\format{
A data frame with 960 rows and 3 variables:
\describe{
  \item{ID}{Character string which refers to the identifier of the gauging station
      (in Austria: HZBCODE)}
  \item{Time}{Character string with date-time information of stage measurements
      which needs to be converted to a compatible format (see \code{\link[=flow]{flow()}})}
  \item{Q}{Flow, stage measurements in \eqn{m^3/s}}
}
}
\usage{
Q
}
\description{
A dataset containing sub-daily flow fluctuations of five consecutive days from
two different gauging stations. One time step is 15 minutes.
}
\keyword{datasets}
